chore(imports): updates imports because of move of common component
